incusd/scriptlet: Refactor load package
